new changes

+++ application/views/member/total_members.php
@@ -8,15 +8,14 @@
                     <thead>
                         <tr>
                             <th>S.No</th>
-                            <th class="th-sm">Full Name</th>
+                            <th class="th-sm">Agent-Id</th>
+                            <th class="th-sm">Agent Name</th>
                             <th class="th-sm">Email Id</th>
                             <th class="th-sm">Phone Number</th>
-                            <!-- <th class="th-sm">Alternate Phone Number</th> -->
                             <th class="th-sm">Gender</th>
-                            <!-- <th class="th-sm">Birth Date</th> -->
-                            <th class="th-sm">Role</th>
-                            <th class="th-sm">Date of Joining</th>
-                            <!-- <th class="th-sm">Other Information</th> -->
+                            <th class="th-sm">Permanent Address</th>
+                            <th class="th-sm">Joining Date</th>
+                            <th class="th-sm">Resignation Date</th>
                             <th class="th-sm">Action</th>
                         </tr>
                     </thead>
@@ -27,23 +26,36 @@
                                     $i++; 
                         ?>
                     <tbody class="table-bordered">
-                        <tr>
-                            <td><?=$i;?></td>
-                            <td><?=$totalmember['name']?></td>
-                            <td><?=$totalmember['email']?></td>
-                            <td><?=$totalmember['phone']?></td>
-                            <!-- <td><?//=$totalmember['alt_phone']?></td> -->
-                            <td><?=$totalmember['gender']?></td>
-                            <!-- <td><?//=$totalmember['dob']?></td> -->
-                            <td><?=$totalmember['role']?></td>
-                            <td><?=$totalmember['joining_date']?></td>
-                            <!-- <td><?//=$totalmember['other_info']?></td> -->
-                            <td class="edit-icon">
-                                <a href="<?=base_url('index.php/member/update_member/'.$totalmember['id'])?>"  class="fa fa-pencil-square-o" title="Edit" data-toggle="modal" aria-hidden="true"></a>&nbsp
-                                <a href="#" onClick = "softDelete(<?=$totalmember['id'];?>);" name="delete" data-href="<?php echo base_url();?>index.php/member/delete_member_soft_data/<?=$totalmember['id'];?>" title="Soft delete" id="delete-<?=$i?>" class="fa fa-trash soft-recode" aria-hidden="true"></a>
-                                <!-- <a href="#" onClick = "hardDelete(<?//=$totalmember['id'];?>);" name="delete" data-href="<?php echo base_url();?>index.php/member/delete_member_hard_data/<?//=$totalmember['id'];?>" title="Hard delete" id="delete-<?//=$i?>" class="fa fa-trash hard-recode" aria-hidden="true"></a> -->
-                            </td>
-                        </tr>
+                        <?php if($totalmember['status'] == 1) { ?>
+                            <tr class="active_agent">
+                                <td><?=$i;?></td>
+                                <td><?="Emp-".$totalmember['id']?></td>
+                                <td><a href="<?=base_url('index.php/member/agent_profile_details/'.$totalmember['id']);?>"><?=$totalmember['name']?></a></td>
+                                <td><?=$totalmember['email']?></td>
+                                <td><?=$totalmember['phone']?></td>
+                                <td><?=$totalmember['gender']?></td>
+                                <td><?=$totalmember['permanent']?></td>
+                                <td><?=$totalmember['joining_date']?></td>
+                                <td></td>
+                                <td class="edit-icon">
+                                    <a href="<?=base_url('index.php/member/update_member/'.$totalmember['id']);?>"  class="fa fa-pencil-square-o" title="Edit" data-toggle="modal" aria-hidden="true"></a>&nbsp
+                                    <a href="#" onClick = "softDelete(<?=$totalmember['id'];?>);" name="delete" data-href="<?php echo base_url();?>index.php/member/delete_member_soft_data/<?=$totalmember['id'];?>" title="Soft delete" id="delete-<?=$i?>" class="fa fa-trash soft-recode" aria-hidden="true"></a>
+                                </td>
+                            </tr>
+                        <?php }else{ ?>
+                            <tr class="inactive_agent">
+                                <td><?=$i;?></td>
+                                <td><?="Emp-".$totalmember['id']?></td>
+                                <td><a href="<?=base_url('index.php/member/agent_profile_details/'.$totalmember['id']);?>"><?=$totalmember['name']?></a></td>
+                                <td><?=$totalmember['email']?></td>
+                                <td><?=$totalmember['phone']?></td>
+                                <td><?=$totalmember['gender']?></td>
+                                <td><?=$totalmember['permanent']?></td>
+                                <td><?=$totalmember['joining_date']?></td>
+                                <td><?=$totalmember['resignation_date']?></td>
+                                <td></td>
+                            </tr>
+                        <?php } ?>    
                     </tbody>
                     <?php } ?>
                     <?php } ?>
